stack output fetching
